requirements.txt
-e .
aws-cdk.core
aws-cdk.aws_s3
aws-cdk.aws-ec2
